+++ dotNET/pdfclown.lib/src/org/pdfclown/documents/contents/colorSpaces/DeviceCMYKColorSpace.cs
@@ -96,7 +96,7 @@ namespace org.pdfclown.documents.contents.colorSpaces
         NOTE: This convertion algorithm was from Apache FOP.
       */
       //FIXME: verify whether this algorithm is effective (limit checking seems quite ugly to me!).
-      float keyCorrection = spaceColor.K / 2.5f;
+      float keyCorrection = (float)spaceColor.K / 2.5f;
       int r = (int)Math.Round((1 - spaceColor.C + keyCorrection) * 255); if(r > 255){r=255;} else if(r < 0){r=0;}
       int g = (int)Math.Round((1 - spaceColor.M + keyCorrection) * 255); if(g > 255){g=255;} else if(g < 0){g=0;}
       int b = (int)Math.Round((1 - spaceColor.Y + keyCorrection) * 255); if(b > 255){b=255;} else if(b < 0){b=0;}
