+++ dotNET/pdfclown.lib/src/org/pdfclown/tools/TextExtractor.cs
@@ -690,7 +690,7 @@ namespace org.pdfclown.tools
         {textStrings.Add(textString = new TextString());}
 
         textStyle = rawTextString.Style;
-        float spaceWidth = textStyle.Font.GetWidth(' ', textStyle.FontSize);
+        double spaceWidth = textStyle.Font.GetWidth(' ', textStyle.FontSize);
         if(spaceWidth == 0)
         {spaceWidth = textStyle.FontSize * .25f;} // NOTE: as a rule of thumb, space width is estimated according to the font size.
         foreach(TextChar textChar in rawTextString.TextChars)
